# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MOVE Apayao — Transit & Trip Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers/location:
    put:
      operationId: updateDriverLocation
      tags: [drivers]
      summary: Push driver GPS location (driver only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationBody"
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /drivers/nearby:
    get:
      operationId: getNearbyDrivers
      tags: [drivers]
      summary: Get nearby available drivers
      security: [{ bearerAuth: [] }]
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lon
          in: query
          required: true
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
            default: 5
      responses:
        "200":
          description: Nearby drivers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearbyDriversResponse"
  /drivers/availability:
    put:
      operationId: updateDriverAvailability
      tags: [drivers]
      summary: Toggle driver availability (driver only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityBody"
      responses:
        "200":
          description: Availability updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfileResponse"
  /drivers/me:
    get:
      operationId: getMyDriverProfile
      tags: [drivers]
      summary: Get own driver profile (driver only)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfileResponse"

  # ── Trips ────────────────────────────────────────────────────────────────────
...
  /incidents:
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Submit an incident report
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentBody"
      responses:
        "201":
          description: Incident reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentResponse"
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Incident list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListIncidentsResponse"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident details
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update incident status (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncidentBody"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentResponse"

  # ── Analytics ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]

    # ── Users / Auth ─────────────────────────────────────────────────────────
...
    DriverProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
          format: uuid
        vehicleType:
          type: string
        licenseNumber:
          type: string
          nullable: true
        plateNumber:
          type: string
          nullable: true
        vehicleColor:
          type: string
          nullable: true
        isAvailable:
          type: boolean
        currentLat:
          type: number
          nullable: true
        currentLon:
          type: number
          nullable: true
        rating:
          type: number
        totalTrips:
          type: integer
      required: [id, userId, vehicleType, isAvailable, rating, totalTrips]
    DriverProfileResponse:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/DriverProfile"
      required: [profile]
    UpdateLocationBody:
      type: object
      properties:
        lat:
          type: number
        lon:
          type: number
      required: [lat, lon]
    AvailabilityBody:
      type: object
      properties:
        isAvailable:
          type: boolean
      required: [isAvailable]
    NearbyDriver:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        vehicleType:
          type: string
        vehicleColor:
          type: string
          nullable: true
        plateNumber:
          type: string
          nullable: true
        isAvailable:
          type: boolean
        currentLat:
          type: number
          nullable: true
        currentLon:
          type: number
          nullable: true
        rating:
          type: number
        totalTrips:
          type: integer
        fullName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        distanceKm:
          type: number
      required: [userId, vehicleType, isAvailable, rating, totalTrips, distanceKm]
    NearbyDriversResponse:
      type: object
      properties:
        drivers:
          type: array
          items:
            $ref: "#/components/schemas/NearbyDriver"
      required: [drivers]

    # ── Trips ────────────────────────────────────────────────────────────────
...
    IncidentReport:
      type: object
      properties:
        id:
          type: integer
        reporterId:
          type: string
          format: uuid
        type:
          type: string
          enum: [accident, flood, fleet_issue]
        lat:
          type: number
        lon:
          type: number
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, reviewing, resolved]
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        resolvedBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, reporterId, type, lat, lon, severity, status, createdAt]
    IncidentResponse:
      type: object
      properties:
        report:
          $ref: "#/components/schemas/IncidentReport"
      required: [report]
    ListIncidentsResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/IncidentReport"
        limit:
          type: integer
        offset:
          type: integer
      required: [reports, limit, offset]
    CreateIncidentBody:
      type: object
      properties:
        type:
          type: string
          enum: [accident, flood, fleet_issue]
        lat:
          type: number
        lon:
          type: number
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
        photoUrl:
          type: string
          format: uri
      required: [type, lat, lon, severity]
    UpdateIncidentBody:
      type: object
      properties:
        status:
          type: string
          enum: [reviewing, resolved]
      required: [status]

    # ── Analytics ────────────────────────────────────────────────────────────
...